pyproject.toml
[tool.poetry]
name = "gmark"
version = "0.2.0"
description = "AI-powered bookmark manager with filesystem-like organization and multi-AI support (Chrome AI, AnythingLLM, OpenAI)"
authors = ["afeldman <anton.feldmann@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
mypy = "^1.13.0"
openai = "^1.54.4"
pydantic = "^2.9.2"
uv = "^0.5.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
decouple = "^0.0.7"
fastapi = "^0.115.5"
fastapi-utils = "^0.7.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
uvicorn = "^0.32.0"
loguru = "^0.7.2"
peewee = "^3.17.8"
fire = "^0.7.0"
tqdm = ">=4.48.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
